tests: keep the packed fixtures portable and their link pairing x86-64
`unsigned long` is 32 bits on Windows, so the alignment checks truncated the
pointer they cast -- the low bits an alignment test reads survive, but Clang
warns and the sizes the callee returns would not. `unsigned long long` is 64
bits on every target here.

The cross-linked pair is held to the same host/target guard the x87 aggregate
pair uses. What the pairing depends on is -fno-pic and this linker's reach
across another compiler's objects, not the layout rules, which are
target-independent -- and the one-file fixture, which carries every size,
alignment and offset, still runs everywhere.

+++ tests/basic_c_packed_layout.c
@@ -161,10 +161,10 @@ int main(void)
     // Objects the linker placed, and one automatic object.
     char automatic_aligned[3] __attribute__((aligned(64)));
     static char static_aligned[3] __attribute__((aligned(32)));
-    if ((unsigned long)(void *)leading_aligned % 64) return 22;
-    if ((unsigned long)(void *)specifier_aligned % 128) return 23;
-    if ((unsigned long)(void *)automatic_aligned % 64) return 24;
-    if ((unsigned long)(void *)static_aligned % 32) return 25;
+    if ((unsigned long long)(void *)leading_aligned % 64) return 22;
+    if ((unsigned long long)(void *)specifier_aligned % 128) return 23;
+    if ((unsigned long long)(void *)automatic_aligned % 64) return 24;
+    if ((unsigned long long)(void *)static_aligned % 32) return 25;
     if (between_them != 1) return 26;
 
     automatic_aligned[0] = 1;
